temp

+++ src/models/track/track.model.js
@@ -1,3 +1,4 @@
+import AppError from "../../utils/appError.js";
 import catchModelAsyncError from "../../utils/catchModelAsyncError.js";
 import handleMongooseError from "../../utils/handleMongooseError.js";
 import Track from "./track.mongo.js";
@@ -18,6 +19,20 @@ export const createTrack = catchModelAsyncError(async (trackData) => {
 }, handleMongooseError);
 
 export const deleteTrack = catchModelAsyncError(async (trackId) => {
+  if (!(await trackExists(trackId))) {
+    throw new AppError("track not found", 404);
+  }
+
   return await Track.findByIdAndDelete(trackId);
 }, handleMongooseError);
 
+export const removePlaylistFromTrack = catchModelAsyncError(async (trackId) => {
+  if (!(await trackExists(trackId))) {
+    throw new AppError("track not found", 404);
+  }
+
+  return await Track.updateMany(
+    { _id: { $in: trackIds } },
+    { $pull: { playlists: playlistId } }
+  );
+}, handleMongooseError);
